backend/requirements.txt
Django==4.0
PyJWT==2.5.0
PyMySQL==1.0.2
